+++ src/locations.rs
@@ -7,6 +7,7 @@ use std::path::{Path, PathBuf};
 const SETTINGS_FILENAME: &str = "version-manager.json";
 
 pub struct Locations {
+    data_local_dir: PathBuf,
     versions_dir: PathBuf,
     config_dir: PathBuf,
 }
@@ -27,6 +28,9 @@ impl Locations {
     pub fn config_dir(&self) -> &Path {
         &self.config_dir
     }
+    pub fn data_local_dir(&self) -> &Path {
+        &self.data_local_dir
+    }
 
     pub fn settings_path(&self) -> PathBuf {
         self.config_dir.join(SETTINGS_FILENAME)
@@ -37,10 +41,12 @@ impl Locations {
     pub fn new() -> Result<Self, NoHomeDirectoryError> {
         let project_dirs =
             ProjectDirs::from("org", "dfinity", "dfx").ok_or(NoHomeDirectoryError)?;
-        let versions_dir = project_dirs.data_local_dir().join("versions");
+        let data_local_dir = project_dirs.data_local_dir().to_path_buf();
+        let versions_dir = data_local_dir.join("versions");
         #[cfg(unix)]
         let config_dir = home_dir()?.join(".config").join("dfx");
         Ok(Self {
+            data_local_dir,
             versions_dir,
             config_dir,
         })