+++ tests/suite/main.rs
@@ -1,4 +1,5 @@
 mod common;
 mod dfx;
 mod dfxvm;
+mod dfxvm_init;
 mod dispatch;